Updated return type of member function 'OnAt' from std::uint8_t to std::size_t

+++ include/bytes/trait/Element.h
@@ -29,7 +29,7 @@ public:
     virtual inline Element & Assign(const Element & cpy) = 0;
     virtual inline Element & Assign(Element && mov) = 0;
 public:
-    virtual inline std::uint8_t OnAt(const std::size_t & i, 
+    virtual inline std::size_t OnAt(const std::size_t & i, 
         const std::size_t & bg, const std::size_t & ed) const = 0;
 public:
     virtual inline bool OnIsEnd(const std::size_t & i,
